# openapi.json
{"openapi":"3.1.0","info":{"title":"Agent Runner","version":"0.1.0"},"paths":{"/v1/chat/completions":{"post":{"tags":["chat"],"summary":"Chat Completions","operationId":"chat_completions_v1_chat_completions_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/health":{"get":{"tags":["admin"],"summary":"Health Check","operationId":"health_check_admin_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/ingestion/status":{"get":{"tags":["admin"],"summary":"Get Ingestion Status","description":"Return status of the RAG ingestion pipeline.","operationId":"get_ingestion_status_admin_admin_ingestion_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/ingestion/resume":{"post":{"tags":["admin"],"summary":"Resume Ingestion","description":"Resume a paused ingestion pipeline.","operationId":"resume_ingestion_admin_admin_ingestion_resume_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/ingestion/pause":{"post":{"tags":["admin"],"summary":"Pause Ingestion","description":"Manually pause the ingestion pipeline.","operationId":"pause_ingestion_admin_admin_ingestion_pause_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/ingestion/clear-and-resume":{"post":{"tags":["admin"],"summary":"Clear And Resume Ingestion","description":"Delete the problem file that caused the pause and resume.","operationId":"clear_and_resume_ingestion_admin_admin_ingestion_clear_and_resume_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/memory/status":{"get":{"tags":["admin"],"summary":"Get Memory Status","description":"Detailed status of the Memory Engine (SurrealDB).","operationId":"get_memory_status_admin_admin_memory_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/health/detailed":{"get":{"tags":["admin"],"summary":"Health Check Detailed","description":"Return a comprehensive health report including topology.","operationId":"health_check_detailed_admin_admin_health_detailed_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/metrics":{"get":{"tags":["admin"],"summary":"Metrics","operationId":"metrics_admin_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/startup-status":{"get":{"tags":["admin"],"summary":"Startup Status","description":"Get the startup status message that was generated during system initialization.\nUseful when no chat client was connected during startup.","operationId":"startup_status_admin_admin_startup_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/system-status":{"get":{"tags":["admin"],"summary":"System Status","operationId":"system_status_admin_admin_system_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/config/models":{"post":{"tags":["admin"],"summary":"Update Models","description":"Update system models dynamically.","operationId":"update_models_admin_admin_config_models_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Config"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/config/update":{"post":{"tags":["admin"],"summary":"Update Single Config","description":"Update a single configuration value or secret.\nBody: { \"key\": \"VISION_MODEL\", \"value\": \"...\", \"type\": \"config\"|\"secret\" }","operationId":"update_single_config_admin_admin_config_update_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/tasks/consolidation":{"post":{"tags":["admin"],"summary":"Trigger Consolidation","description":"Manually trigger memory consolidation.","operationId":"trigger_consolidation_admin_admin_tasks_consolidation_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/tasks/backup":{"post":{"tags":["admin"],"summary":"Trigger Backup","description":"Manually trigger memory backup.","operationId":"trigger_backup_admin_admin_tasks_backup_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/circuit-breaker/status":{"get":{"tags":["admin"],"summary":"Circuit Breaker Status","operationId":"circuit_breaker_status_admin_admin_circuit_breaker_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/circuit-breaker/reset/{name}":{"post":{"tags":["admin"],"summary":"Reset Circuit Breaker","operationId":"reset_circuit_breaker_admin_admin_circuit_breaker_reset__name__post","parameters":[{"name":"name","in":"path","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/circuit-breaker/reset-all":{"post":{"tags":["admin"],"summary":"Reset All Circuit Breakers","operationId":"reset_all_circuit_breakers_admin_admin_circuit_breaker_reset_all_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/telemetry/log":{"post":{"tags":["admin"],"summary":"Log Telemetry","description":"Receive logs/errors from the frontend dashboard.","operationId":"log_telemetry_admin_admin_telemetry_log_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/background-tasks":{"get":{"tags":["admin"],"summary":"List Tasks","operationId":"list_tasks_admin_admin_background_tasks_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/system-prompt":{"get":{"tags":["admin"],"summary":"Get System Prompt","operationId":"get_system_prompt_admin_admin_system_prompt_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/memory/facts":{"get":{"tags":["admin"],"summary":"Get Memory Facts","operationId":"get_memory_facts_admin_admin_memory_facts_get","parameters":[{"name":"query","in":"query","required":false,"schema":{"type":"string","default":"","title":"Query"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/dashboard/insights":{"get":{"tags":["admin"],"summary":"Get Dashboard Insights","operationId":"get_dashboard_insights_admin_admin_dashboard_insights_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/dashboard/state":{"get":{"tags":["admin"],"summary":"Get Dashboard State","description":"Aggregate system state for the V2 Dashboard.","operationId":"get_dashboard_state_admin_admin_dashboard_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/dashboard/errors":{"post":{"tags":["admin"],"summary":"Track Dashboard Error","operationId":"track_dashboard_error_admin_admin_dashboard_errors_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/dashboard/interactions":{"post":{"tags":["admin"],"summary":"Track Dashboard Interaction","operationId":"track_dashboard_interaction_admin_admin_dashboard_interactions_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/system/logs":{"get":{"tags":["admin"],"summary":"Get Logs Tail","description":"Return the last N lines of the agent runner log.","operationId":"get_logs_tail_admin_admin_system_logs_get","parameters":[{"name":"lines","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Lines"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/llm/roles":{"get":{"tags":["admin"],"summary":"Get Llm Roles","operationId":"get_llm_roles_admin_admin_llm_roles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}},"post":{"tags":["admin"],"summary":"Update Llm Roles","operationId":"update_llm_roles_admin_admin_llm_roles_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/system/process/stop":{"post":{"tags":["admin"],"summary":"Stop Process","description":"Stop the Agent Runner process.","operationId":"stop_process_admin_admin_system_process_stop_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/mcp/server/status":{"get":{"tags":["admin"],"summary":"Get Mcp Server Status","description":"Get status of the internal MCP Server (SSE).","operationId":"get_mcp_server_status_admin_admin_mcp_server_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/logs/stream":{"get":{"tags":["admin"],"summary":"Stream Logs","description":"Stream logs via SSE using native StreamingResponse.\nTail relevant log files and emit lines as they are written.","operationId":"stream_logs_admin_admin_logs_stream_get","parameters":[{"name":"services","in":"query","required":false,"schema":{"type":"string","default":"agent_runner","title":"Services"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/system/restart":{"post":{"tags":["admin"],"summary":"Restart System Endpoint","description":"Trigger a full system restart script.","operationId":"restart_system_endpoint_admin_admin_system_restart_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/diagnostics/stream":{"get":{"tags":["admin"],"summary":"Stream Diagnostics","description":"Stream diagnostic logs directly from the diagnostic_log table.\n\nNOTE: This endpoint is currently UNUSED - there is no diagnostics tab in the UI.\nlive_stream.md is a log file only - it should not be read for UI purposes.\nIf a diagnostics UI is added in the future, this endpoint is ready to use.","operationId":"stream_diagnostics_admin_admin_diagnostics_stream_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/tools/execute":{"post":{"tags":["admin"],"summary":"Execute Tool Endpoint","description":"Directly execute a defined tool/function.\nBody: { \"tool_name\": \"list_dir\", \"arguments\": {\"path\": \".\"} }","operationId":"execute_tool_endpoint_admin_admin_tools_execute_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Body"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/admin/admin/registry/health":{"get":{"tags":["admin"],"summary":"Registry Health","description":"Check registry integrity and return validation status.","operationId":"registry_health_admin_admin_registry_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/task-health":{"get":{"tags":["admin"],"summary":"Task Health","description":"Get health statistics for fire-and-forget tasks.","operationId":"task_health_admin_admin_task_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/validate-changes":{"post":{"tags":["admin"],"summary":"Validate Changes","description":"Validate all recent changes (last 2 hours).\nTests tool categorization, evaluation, registry, rating integration, and marketplace.","operationId":"validate_changes_admin_admin_validate_changes_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/admin/admin/chat-latency-test":{"post":{"tags":["admin"],"summary":"Chat Latency Test","description":"Run comprehensive chat interface latency test with component breakdown.","operationId":"chat_latency_test_admin_admin_chat_latency_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_chat_latency_test_admin_admin_chat_latency_test_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/reload-mcp":{"post":{"tags":["mcp"],"summary":"Reload Mcp","operationId":"reload_mcp_mcp_admin_mcp_reload_mcp_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/mcp/admin/mcp/add":{"post":{"tags":["mcp"],"summary":"Add Mcp Server Endpoint","description":"Dynamically add or update an MCP server and discover its tools.","operationId":"add_mcp_server_endpoint_mcp_admin_mcp_add_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_add_mcp_server_endpoint_mcp_admin_mcp_add_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/install":{"post":{"tags":["mcp"],"summary":"Install Mcp Package Endpoint","description":"Install an NPM-based MCP server package automatically.","operationId":"install_mcp_package_endpoint_mcp_admin_mcp_install_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_install_mcp_package_endpoint_mcp_admin_mcp_install_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/remove":{"post":{"tags":["mcp"],"summary":"Remove Mcp Server Endpoint","description":"Dynamically remove an MCP server, terminate its process, and clear tool cache.","operationId":"remove_mcp_server_endpoint_mcp_admin_mcp_remove_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_remove_mcp_server_endpoint_mcp_admin_mcp_remove_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/toggle":{"post":{"tags":["mcp"],"summary":"Toggle Mcp Endpoint","description":"Dynamically enable or disable an MCP server.","operationId":"toggle_mcp_endpoint_mcp_admin_mcp_toggle_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_toggle_mcp_endpoint_mcp_admin_mcp_toggle_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/upload-config-json":{"post":{"tags":["mcp"],"summary":"Upload Mcp Config Json","description":"JSON-only version of upload-config for easier proxying.","operationId":"upload_mcp_config_json_mcp_admin_mcp_upload_config_json_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/upload-config":{"post":{"tags":["mcp"],"summary":"Upload Mcp Config","description":"LLM-powered endpoint to parse and add MCP servers from raw text/files.","operationId":"upload_mcp_config_mcp_admin_mcp_upload_config_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_mcp_config_mcp_admin_mcp_upload_config_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/tools":{"get":{"tags":["mcp"],"summary":"Get All Mcp Tools","description":"Return discovered MCP tools, optionally filtered by server.","operationId":"get_all_mcp_tools_mcp_admin_mcp_tools_get","parameters":[{"name":"server","in":"query","required":false,"schema":{"type":"string","title":"Server"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/mcp/admin/mcp/debug/call":{"post":{"tags":["mcp"],"summary":"Debug Mcp Call","description":"Directly execute an MCP tool via the proxy for debugging.","operationId":"debug_mcp_call_mcp_admin_mcp_debug_call_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_debug_mcp_call_mcp_admin_mcp_debug_call_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/system/mcp/sse":{"get":{"tags":["system"],"summary":"Mcp Sse Endpoint","description":"Standard MCP SSE Endpoint.\nEstablishes a connection and yields JSON-RPC events.","operationId":"mcp_sse_endpoint_system_mcp_sse_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/system/mcp/messages":{"post":{"tags":["system"],"summary":"Mcp Messages Endpoint","description":"Handle incoming JSON-RPC messages from the client.\nNote: MCP uses a separate POST endpoint for client -> server messages.","operationId":"mcp_messages_endpoint_system_mcp_messages_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/files/admin/upload":{"post":{"tags":["files"],"summary":"Upload File Endpoint","description":"Handle uploads from the dashboard (Paste-to-Upload).\nAccepts specific file uploads OR raw text content.\nSaves to WORKSPACE_ROOT/uploads/.","operationId":"upload_file_endpoint_files_admin_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_endpoint_files_admin_upload_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/files/admin/docs/list":{"get":{"tags":["files"],"summary":"List Documentation Files","description":"List markdown files in the ai/docs directory.","operationId":"list_documentation_files_files_admin_docs_list_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/files/admin/docs/read":{"get":{"tags":["files"],"summary":"Read Documentation File","operationId":"read_documentation_file_files_admin_docs_read_get","parameters":[{"name":"path","in":"query","required":true,"schema":{"type":"string","title":"Path"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_add_mcp_server_endpoint_mcp_admin_mcp_add_post":{"properties":{"name":{"type":"string","title":"Name"},"config":{"additionalProperties":true,"type":"object","title":"Config"}},"type":"object","required":["name","config"],"title":"Body_add_mcp_server_endpoint_mcp_admin_mcp_add_post"},"Body_chat_latency_test_admin_admin_chat_latency_test_post":{"properties":{"iterations":{"type":"integer","title":"Iterations","default":5},"test_message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Test Message"}},"type":"object","title":"Body_chat_latency_test_admin_admin_chat_latency_test_post"},"Body_debug_mcp_call_mcp_admin_mcp_debug_call_post":{"properties":{"server":{"type":"string","title":"Server"},"tool":{"type":"string","title":"Tool"},"arguments":{"additionalProperties":true,"type":"object","title":"Arguments","default":{}}},"type":"object","required":["server","tool"],"title":"Body_debug_mcp_call_mcp_admin_mcp_debug_call_post"},"Body_install_mcp_package_endpoint_mcp_admin_mcp_install_post":{"properties":{"package":{"type":"string","title":"Package"},"name":{"type":"string","title":"Name"},"args":{"items":{"type":"string"},"type":"array","title":"Args"}},"type":"object","required":["package"],"title":"Body_install_mcp_package_endpoint_mcp_admin_mcp_install_post"},"Body_remove_mcp_server_endpoint_mcp_admin_mcp_remove_post":{"properties":{"name":{"type":"string","title":"Name"}},"type":"object","required":["name"],"title":"Body_remove_mcp_server_endpoint_mcp_admin_mcp_remove_post"},"Body_toggle_mcp_endpoint_mcp_admin_mcp_toggle_post":{"properties":{"name":{"type":"string","title":"Name"},"enabled":{"type":"boolean","title":"Enabled"}},"type":"object","required":["name","enabled"],"title":"Body_toggle_mcp_endpoint_mcp_admin_mcp_toggle_post"},"Body_upload_file_endpoint_files_admin_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"content":{"type":"string","title":"Content"},"filename":{"type":"string","title":"Filename"}},"type":"object","title":"Body_upload_file_endpoint_files_admin_upload_post"},"Body_upload_mcp_config_mcp_admin_mcp_upload_config_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"raw_text":{"type":"string","title":"Raw Text"}},"type":"object","title":"Body_upload_mcp_config_mcp_admin_mcp_upload_config_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}